+++ src/snowflake/snowpark/_internal/data_source/drivers/base_driver.py
@@ -98,11 +98,7 @@ class BaseDriver:
             select * from {partition_table}, table({udtf_name}({PARTITION_TABLE_COLUMN_NAME}))
             """
         res = session.sql(call_udtf_sql)
-        cols = [
-            res[field.name].cast(field.datatype).alias(field.name)
-            for field in schema.fields
-        ]
-        return res.select(cols)
+        return self.to_result_snowpark_df_udtf(res, schema)
 
     def udtf_class_builder(self, fetch_size: int = 1000) -> type:
         create_connection = self.create_connection
@@ -169,3 +165,15 @@ class BaseDriver:
         session: "Session", table_name, schema, _emit_ast: bool = True
     ) -> "DataFrame":
         return session.table(table_name, _emit_ast=_emit_ast)
+
+    @staticmethod
+    def to_result_snowpark_df_udtf(
+        res_df: "DataFrame",
+        schema: StructType,
+        _emit_ast: bool = True,
+    ):
+        cols = [
+            res_df[field.name].cast(field.datatype).alias(field.name)
+            for field in schema.fields
+        ]
+        return res_df.select(cols, _emit_ast=_emit_ast)